Refactor login page containers into components: LoginForm, LoginCard, AuthFooter, SocialLogin, LoadingOverlay, and TopNavBar; integrate theme context for styling
